FormPengembalian kurang delete dan pilih data untuk diedit

+++ WindowsFormsApp1/formUser.cs
@@ -13,7 +13,7 @@ namespace WindowsFormsApp1
 {
     public partial class formUser : Form
     {
-        private int SelectedUserID;
+        private int SelectedUserID = 0;
         public formUser()
         {
             InitializeComponent();
@@ -23,7 +23,7 @@ namespace WindowsFormsApp1
 
         private void ClearFields()
         {
-            txtUsername.Clear();
+            txtEmail.Clear();
             txtPassword.Clear();
             SelectedUserID = 0; 
         }
@@ -37,7 +37,7 @@ namespace WindowsFormsApp1
                 try
                 {
                     conn.Open();
-                    string sql = "SELECT Users.id, Users.username, Users.password, [Status].nama_status" +
+                    string sql = "SELECT Users.id, Users.email, Users.password, [Status].nama_status" +
                                   "  FROM Users JOIN [Status] ON Users.status_id = [status].id";
 
                     SqlDataAdapter da = new SqlDataAdapter(sql, conn);
@@ -79,7 +79,7 @@ namespace WindowsFormsApp1
         }
 
 
-        //memunculkan pilihan status!
+        //memunculkan ComboBox status!
         private void LoadStatus()
         {
             using (SqlConnection conn = Classkoneksi.GetConn())
@@ -98,6 +98,9 @@ namespace WindowsFormsApp1
                     cmbStatus.DataSource = dt;
                     cmbStatus.DisplayMember = "nama_status";
                     cmbStatus.ValueMember = "id";
+
+                    //Agar Cmb kosong saat pertama kali form dibuka!
+                    cmbStatus.SelectedIndex = 0;
                 }
                 catch (Exception ex)
                 {
@@ -112,7 +115,7 @@ namespace WindowsFormsApp1
         //Tombol create Users 
         private void btntambah_Click(object sender, EventArgs e)
         {
-            if (string.IsNullOrEmpty(txtUsername.Text) || string.IsNullOrEmpty(txtPassword.Text)) 
+            if (string.IsNullOrEmpty(txtEmail.Text) || string.IsNullOrEmpty(txtPassword.Text)) 
             {
                 MessageBox.Show("Semua Kolom Wajib Diisi!");
                 return;
@@ -123,10 +126,10 @@ namespace WindowsFormsApp1
                 try
                 {
                     conn.Open();
-                    string sql = "INSERT INTO Users (username, password, status_id) VALUES (@username, @password, @status) ";
+                    string sql = "INSERT INTO Users (email, password, status_id) VALUES (@email, @password, @status) ";
                     SqlCommand cmd = new SqlCommand(sql, conn);
 
-                    cmd.Parameters.AddWithValue("@username", txtUsername.Text);
+                    cmd.Parameters.AddWithValue("@email", txtEmail.Text);
                     cmd.Parameters.AddWithValue("@password", txtPassword.Text);
                     cmd.Parameters.AddWithValue("@status", cmbStatus.SelectedValue);
 
@@ -162,11 +165,11 @@ namespace WindowsFormsApp1
                 {
                     conn.Open();
 
-                    string sql = "UPDATE Users SET username=@username, password=@password, status_id=@status WHERE id=@id ";
+                    string sql = "UPDATE Users SET email=@email, password=@password, status_id=@status WHERE id=@id ";
                     SqlCommand cmd = new SqlCommand(sql, conn);
 
                     cmd.Parameters.AddWithValue("@id", SelectedUserID);
-                    cmd.Parameters.AddWithValue("@username", txtUsername.Text);
+                    cmd.Parameters.AddWithValue("@email", txtEmail.Text);
                     cmd.Parameters.AddWithValue("@password", txtPassword.Text);
                     cmd.Parameters.AddWithValue("@status", cmbStatus.SelectedValue);
 
@@ -189,31 +192,39 @@ namespace WindowsFormsApp1
         private void btnhapus_Click(object sender, EventArgs e)
 
         {
+            
             if (SelectedUserID == 0)
             {
                 MessageBox.Show("Pilih Data yang ingin dihapus!");
                 return;
             }
 
-            using (SqlConnection conn = Classkoneksi.GetConn())
-                try
-                {
-                    conn.Open();
-                    string sql = "DELETE FROM Users WHERE id=@id";
-                    SqlCommand cmd = new SqlCommand(sql, conn);
-                    cmd.Parameters.AddWithValue("@id", SelectedUserID);
+            //KOnfirmasi hapus!
 
-                    cmd.ExecuteNonQuery();
-                    MessageBox.Show("User Berhasil Dihapus");
-                    
+            DialogResult dr = MessageBox.Show("Yakin ingin Menghapus User?", "Konfirmasi"  , MessageBoxButtons.YesNo);
+            if (dr == DialogResult.Yes)
+            {
+                using (SqlConnection conn = Classkoneksi.GetConn())
+                    try
+                    {
+                        conn.Open();
+                        string sql = "DELETE FROM Users WHERE id=@id";
+                        SqlCommand cmd = new SqlCommand(sql, conn);
+                        cmd.Parameters.AddWithValue("@id", SelectedUserID);
 
-                    TampilUser();
-                    ClearFields() ;
-                }
-                catch (Exception ex)
-                {
-                    MessageBox.Show(" Gagal menghapus Data User" + ex.Message);
-                }
+                        cmd.ExecuteNonQuery();
+                        MessageBox.Show("User Berhasil Dihapus");
+
+
+                        TampilUser();
+                        ClearFields();
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show(" Gagal menghapus Data User" + ex.Message);
+                    }
+            }
+           
         }
 
 
@@ -224,14 +235,14 @@ namespace WindowsFormsApp1
 
 
 
-//Agar Data muncul pada DGV!!
+//Agar Data DApat dipilih pada DGV!!
 private void dgvUsers_CellClick(object sender, DataGridViewCellEventArgs e)
         {
             if (e.RowIndex >= 0 && dgvUsers.Rows[e.RowIndex].Cells[0].Value != null)
             {
                 DataGridViewRow row = dgvUsers.Rows[e.RowIndex];
                 SelectedUserID = int.Parse(row.Cells[0].Value.ToString());
-                txtUsername.Text = row.Cells[1].Value.ToString();
+                txtEmail.Text = row.Cells[1].Value.ToString();
                 txtPassword.Text = row.Cells[2].Value.ToString();
                 cmbStatus.Text = row.Cells[3].Value.ToString();
